Rename lib files to kebab-case per contributing guidelines
product-filters.ts (was productFilters.ts)
product-sort.ts (was productSort.ts)
